requirements.txt

torchvision==0.2.1
Werkzeug==0.16.0
matplotlib==3.1.2
pandas==0.25.3
Flask==1.1.1
numpy==1.17.4
seaborn==0.9.0
wheel==0.33.6
Pillow==6.2.1
gunicorn==20.0.4
https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
torchvision==0.2.1
scikit-learn==0.23.2
Flask-Cors==3.0.8